backend/requirements.txt
transformers==4.37.2
torch==2.2.0
torchaudio==2.2.0
fastapi==0.109.2
python-multipart==0.0.6
numpy==1.24.3
uvicorn==0.27.1
python-dotenv==1.0.0
soundfile==0.12.1 
librosa
ibm-watson
ibm-cloud-sdk-core
python-dotenv
google-genai